+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 10.12.1 (11-12-2025)
+
+### Bugfix: 2
+- [#39516](https://parermine.regione.emilia-romagna.it/issues/39516) Correzione vista per estrazione strutture coinvolte nella creazione della richiesta di restituzione arichivio.
+- [#39494](https://parermine.regione.emilia-romagna.it/issues/39494) Correzione metodo di generazione file in fase di esportazione di una ricerca
 
 ## 10.12.0 (21-11-2025)